Fix printf format for Point coordinates in debug output
- Cast SInt16 coordinates to int for proper printf formatting
- This fixes debug output but doesn't solve the underlying click handling issues

+++ src/EventManager/EventDispatcher.c
@@ -135,7 +135,7 @@ Boolean HandleMouseDown(EventRecord* event)
     windowPart = FindWindow(event->where, &whichWindow);
 
     serial_printf("HandleMouseDown: part=%d, window=%p at (%d,%d)\n",
-                 windowPart, whichWindow, event->where.h, event->where.v);
+                 windowPart, whichWindow, (int)event->where.h, (int)event->where.v);
 
     switch (windowPart) {
         case inMenuBar:
